change np.float to np.float32

+++ genetic/crossover_and_mutation.py
@@ -431,7 +431,7 @@ class Mutation(object):
         idx = np.argsort(a)
         idx = idx[::-1]
         sort_a = a[idx]
-        sum_a = np.sum(a).astype(np.float)
+        sum_a = np.sum(a).astype(np.float32)
         selected_index = []
         for _ in range(k):
             u = np.random.rand() * sum_a
